+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
-## Unreleased
+## [0.7.17] 2020-03-06
+
+### Added
+
+- field completion for CDS views
+- data source completion for CDS views
 
 ## [0.7.16] 2020-02-03